// src/shared/types/utm.ts
// shared/types/utm.ts
export interface UtmParams {
  utm_source?: string;
  utm_medium?: string;
  utm_campaign?: string;
  utm_term?: string;
  utm_content?: string;
  utm_name?: string;
  utm_phone?: string;
  utm_email?: string;
  utm_leadid?: string;
  utm_yclientid?: string;
  utm_gaclientid?: string;
  ref_user?: string;
}